Update tests to use pytch-testing module
Large diff, but fairly mechanical.  The general pattern of the changes
is:

Require the new module, bringing in the functions which were
previously at global scope.  Immediately call the configure_mocha()
function to set up beforeEach() and afterEach() hooks.

Wrap mocha's it() (or sometimes describe()) calls in a with_module()
or with_project() as appropriate.  Inside the it(), use the passed-in
nullary import_module function instead of the previous
import_local_file(), or use the passed-in nullary import_project
function instead of the previous import_project().  Sometimes several
it() calls use the same module or project, and these can all be
brought inside the same with_module() or with_project() call.

When test-cases are specified in a list and then used within a
forEach(), that code is brought into the with_project() call.

Some special cases: The local async_project() for bad registrations
can be eliminated as its job is done by the passed-in import_project
function.

The two_threads_project() function in test_scheduling.js can be
brought into the with_project() body, leaving the it() calls
unchanged.  Similarly for the single_sprite_project() and
sprite_on_stage() functions in test_hat_blocks.js, and the
prepare_project() function in test_instance_discovery.js.