unregister_actor_instance(): Remove
Was used in Sprite.delete_this_clone() but no longer.

It's possible we'll need to put something like this back one day, if
we want the ability to unregister things other than the instance on
which the current thread is running.  But until then, it may as well
not clutter up the code.